Add controller to find client by id

+++ src/main/java/com/insurance/insurance_api/repository/CompanyRepository.java
@@ -3,5 +3,5 @@ package com.insurance.insurance_api.repository;
 import com.insurance.insurance_api.entity.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyRepository extends JpaRepository<Company, Integer> {
+public interface CompanyRepository extends JpaRepository<Company, Long> {
 }
